Finished replacing the REST MangaClient with the GraphQL MangaClient
